CLN Migrate avaliable_if to it's own file
