+++ sklearn/utils/metaestimators.py
@@ -3,19 +3,17 @@
 #         Andreas Mueller
 # License: BSD
 from typing import List, Any
-from types import MethodType
 import warnings
-from functools import wraps
 
 from abc import ABCMeta, abstractmethod
 from operator import attrgetter
-from functools import update_wrapper
 import numpy as np
 from contextlib import suppress
 
 from ..utils import _safe_indexing
 from ..utils._tags import _safe_tags
 from ..base import BaseEstimator
+from ._available_if import available_if, _AvailableIfDescriptor
 
 __all__ = ["available_if", "if_delegate_has_method"]
 
@@ -96,90 +94,6 @@ class _BaseComposition(BaseEstimator, metaclass=ABCMeta):
             raise ValueError(
                 "Estimator names must not contain __: got {0!r}".format(invalid_names)
             )
-
-
-class _AvailableIfDescriptor:
-    """Implements a conditional property using the descriptor protocol.
-
-    Using this class to create a decorator will raise an ``AttributeError``
-    if check(self) returns a falsey value. Note that if check raises an error
-    this will also result in hasattr returning false.
-
-    See https://docs.python.org/3/howto/descriptor.html for an explanation of
-    descriptors.
-    """
-
-    def __init__(self, fn, check, attribute_name):
-        self.fn = fn
-        self.check = check
-        self.attribute_name = attribute_name
-
-        # update the docstring of the descriptor
-        update_wrapper(self, fn)
-
-    def __get__(self, obj, owner=None):
-        attr_err = AttributeError(
-            f"This {repr(owner.__name__)} has no attribute {repr(self.attribute_name)}"
-        )
-        if obj is not None:
-            # delegate only on instances, not the classes.
-            # this is to allow access to the docstrings.
-            if not self.check(obj):
-                raise attr_err
-            out = MethodType(self.fn, obj)
-
-        else:
-            # This makes it possible to use the decorated method as an unbound method,
-            # for instance when monkeypatching.
-            @wraps(self.fn)
-            def out(*args, **kwargs):
-                if not self.check(args[0]):
-                    raise attr_err
-                return self.fn(*args, **kwargs)
-
-        return out
-
-
-def available_if(check):
-    """An attribute that is available only if check returns a truthy value.
-
-    Parameters
-    ----------
-    check : callable
-        When passed the object with the decorated method, this should return
-        a truthy value if the attribute is available, and either return False
-        or raise an AttributeError if not available.
-
-    Returns
-    -------
-    callable
-        Callable makes the decorated method available if `check` returns
-        a truthy value, otherwise the decorated method is unavailable.
-
-    Examples
-    --------
-    >>> from sklearn.utils.metaestimators import available_if
-    >>> class HelloIfEven:
-    ...    def __init__(self, x):
-    ...        self.x = x
-    ...
-    ...    def _x_is_even(self):
-    ...        return self.x % 2 == 0
-    ...
-    ...    @available_if(_x_is_even)
-    ...    def say_hello(self):
-    ...        print("Hello")
-    ...
-    >>> obj = HelloIfEven(1)
-    >>> hasattr(obj, "say_hello")
-    False
-    >>> obj.x = 2
-    >>> hasattr(obj, "say_hello")
-    True
-    >>> obj.say_hello()
-    Hello
-    """
-    return lambda fn: _AvailableIfDescriptor(fn, check, attribute_name=fn.__name__)
 
 
 # TODO(1.3) remove
